requirements.txt
ocrmypdf>=13.6.0
pdf2image>=1.16.0
Pillow>=9.0.0
opencv-python-headless>=4.5.0
rich>=12.0.0
pytesseract>=0.3.8
tqdm>=4.62.0
